requirements.txt
streamlit
google-genai
PyPDF2
gTTS
streamlit-mic-recorder
python-dotenv